tests_perf/requirements.txt
# google-cloud-bigquery == 1.11.2
# grpcio == 1.19.0
# web3 == 4.9.1
ethereum-etl == 1.3.0
pyetherchain == 0.3.2
boto3 == 1.9.132
cython
python-rocksdb == 0.7.0
click
frozendict == 1.2